3_runVEP.pl : don't die if stdin has no data

+++ 3_runVEP.pl
@@ -252,17 +252,19 @@ Arguments (all can be abbreviated to shortest unambiguous prefixes):
 	}
     }
 
-    # merge VCFs for last chrom
-    $now = strftime("%F %T", localtime);
-    warn "I $now: $0 - finished parsing/processing chrom $prevChr\n";
-    close($VCFVEP);
-    close($VCFCACHE);
-    # merge last chr files
-    &mergeAndPrint($vcfFromVep, $vcfFromCache, $cacheUpdate);
-    unlink($vcfFromVep, $vcfFromCache);
+    # merge VCFs for last chrom, unless stdin was completely empty
+    if ($prevChr) {
+	$now = strftime("%F %T", localtime);
+	warn "I $now: $0 - finished parsing/processing chrom $prevChr\n";
+	close($VCFVEP);
+	close($VCFCACHE);
+	# merge last chr files
+	&mergeAndPrint($vcfFromVep, $vcfFromCache, $cacheUpdate);
+	unlink($vcfFromVep, $vcfFromCache);
 
-    # update cache with new CSQs
-    &updateCache($cacheFile, $cacheUpdate, $debug);
+	# update cache with new CSQs
+	&updateCache($cacheFile, $cacheUpdate, $debug);
+    }
 
     # clean up
     unlink($vepStats) || die "E $0: cannot unlink vepStats file $vepStats\n";
